+++ srcs/ft_isdigit.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_isdigit.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fcatusse <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 15:55:26 by fcatusse          #+#    #+#             */
-/*   Updated: 2019/07/19 12:17:14 by fcatusse         ###   ########.fr       */
+/*   Updated: 2020/03/05 01:20:33 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "shared_libft.h"
 
-int		ft_isdigit(int c)
+int16_t	ft_isdigit(int16_t c)
 {
 	return (c >= '0' && c <= '9');
 }
