+++ srcs/ft_strrchr.c
@@ -6,13 +6,13 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 16:12:26 by fcatusse          #+#    #+#             */
-/*   Updated: 2020/03/04 23:13:29 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/05 01:13:09 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "shared_libft.h"
 
-char	*ft_strrchr(const char *s, int c)
+char	*ft_strrchr(const char *s, int16_t c)
 {
 	size_t	i;
